test(imgproc): move histogram module tests into src/imgproc/tests.rs
Every other imgproc submodule (color.rs, edge.rs, filter.rs, morph.rs,
pyramid.rs, threshold.rs, hough.rs, geometric.rs, resize.rs,
derivatives.rs, feature.rs) keeps its tests solely in the shared
src/imgproc/tests.rs, per the module structure convention in
CLAUDE.md. histogram.rs was the only one with its own embedded
#[cfg(test)] mod tests block, left over from how #99 originally wrote
it.

Mechanical move, no behavior change: all 33 histogram tests already
went through the public API (calc_hist, calc_back_project,
compare_hist, equalize_hist, Clahe/create_clahe, RangeSpec,
HistCompMethods) with no dependency on private helpers, so they drop
straight into tests.rs's existing imgproc_tests module (which already
glob-imports crate::core::* and crate::imgproc::*) with no import
changes needed.

Verified: same 342 lib test count before and after (nothing lost or
duplicated), passing under both default and
--no-default-features --features std, fmt/clippy clean under
--all-features and --no-default-features.

Closes #111
